// src/pages/Senaia/IDENTIFICA/IdentificaScreen/types.ts
/**
 // ########################################
// #  www.pydoundation.org
// #  Eng Marcelo Anjos
// #  marcelu.phd@gmail.com
// #  2022-02-10
// ########################################
 * @ Create Time: 2022-04-16 20:57:35
 * @ Modified by: Your name
 * @ Modified time: 2022-04-17 10:42:22
 * @ Description:
 */

// ###############################   TYPES  ###############################
export interface IIdentifica {
  dbversion?: string;
  identificaId?: string;
  identificaUniqueId: string;
  identifica_company?: string;
  identificaIdAnimal?: string;
  identificaNroTag?: string;
  identificaIdTecnico?: string;
  identificaNumeroIdentificador?: string;
  identificaDataNascimento?: string;
  identificaSexo?: string;
  identificaIdRaca?: string;
  identificaMarcaFogo?: string;
  identificaLatitude?: string;
  identificaLongitude?: string;
  identificaCarimbo?: string;
  identificaLoteVacina?: string;
  identificaPeso?: string;
  identificaCodErro?: string;
  identificaDataIdentificacao?: string;
}

export type TIdentifica = {
  dbversion?: string;
  identificaId?: string;
  identificaUniqueId: string;
  identifica_company?: string;
  identificaIdAnimal?: string;
  identificaNroTag?: string;
  identificaIdTecnico?: string;
  identificaNumeroIdentificador?: string;
  identificaDataNascimento?: string;
  identificaSexo?: string;
  identificaIdRaca?: string;
  identificaMarcaFogo?: string;
  identificaLatitude?: string;
  identificaLongitude?: string;
  identificaCarimbo?: string;
  identificaLoteVacina?: string;
  identificaPeso?: string;
  identificaCodErro?: string;
  identificaDataIdentificacao?: string;
};